moved relative quiz Java Script logic to quiz.js file so that the duplication issues of results title and trailer links have been resolved
