+++ .vscode/settings.json
@@ -1,3 +1,3 @@
 {
-    "liveServer.settings.port": 5503
+    "liveServer.settings.port": 5505
 }
