+++ assets/js/quiz.js
@@ -1,7 +1,26 @@
+
+// Starter Variables
+let recommendedMovie = "";
+let movieID = "";
+
+// API URL Variables
+let apiKeyTMBD = "87ceec9af92ce89acfb2e11778f0841f";
+let idURL;
+
+// Variable Elements from HTML
+let trailerSourceEl = document.getElementById('src');
+let trailerVideoEl = document.getElementById('video');
+let trailerEl = document.getElementById('link-to-trailer');
+let videoLink;
+let recommendedTitleEl = document.getElementById("recommended-title");
+// let trailerContainerEl = document.getElementById('trailer-container');
+
+// Variables defined to show & hide HTML elements on page
 let quizContainerEl = $("#quiz-container");
 let trailerContainerEl= $('#link-to-trailer');
 let resultsContainerEl = $("#results");
 let resultsTitleEl = $("#recommended-title");
+
 
 quizContainerEl.css("display", "block");
 resultsContainerEl.css("display", "none");
@@ -85,10 +104,75 @@ $(function () {
 			$("#results").append("<div class='pad-8'>" + overview + "</div><hr class='hr'>");
 			console.log("title" + title);
 			getMovieID (title);
-
-		// getMovieID (results.results[5].title);
 		}
 	}
+
+// Function to call GET Search Movies to get Movie ID
+function getMovieID (title){
+    recommendedMovie = title;
+    console.log(title);
+    idURL= "https://api.themoviedb.org/3/search/movie?api_key=" + apiKeyTMBD + "&language=en-US&query=" + recommendedMovie + "&page=1&include_adult=false";
+    fetch(idURL)
+    .then(function(response){
+        if (response.ok) {
+            // need to look at other examples to make sure the .then below is correct as instructor said this isn't happening
+            response.json().then(function (data) {
+                console.log("API call was a success!");
+                console.log(data.results[0].id);
+                movieID=data.results[0].id;
+                Trailers(movieID);                
+              });
+            }      
+        })
+}
+
+
+
+// Function to populate Trailer with movie ID
+function Trailers (movieID) {
+    let trailerURL = "https://api.themoviedb.org/3/movie/" + movieID + "/videos?api_key=" + apiKeyTMBD + "&language=en-US";
+    // Chelsea's back up plan if trailerURL doesn't work
+    // let videoURL = "https://api.themoviedb.org/3/movie/" + movieID + "?api_key=" + apiKeyTMBD + "&append_to_response=videos,images";
+   
+    fetch(trailerURL)
+    .then(function(response){
+            if (response.ok) {
+                response.json().then(function (data) {
+                    console.log(data);
+                // Pull they video key from the API Trailer Array
+                 let videoKey = data.results[0].key;
+                //  console.log(videoKey);
+                let trailerMovieName =data.results[0].name;
+                console.log(trailerMovieName);
+                //  Create the Youtube Link with the key of the video
+                let YoutubeLink = "https://www.youtube.com/watch?v=" + videoKey;
+                console.log(YoutubeLink);
+
+                // Find organized list element
+                let trailerListEl = document.querySelector("#trailer-list");
+            
+                // Creates trailer Link list Element
+                let li1 = document.createElement("li");
+                // li1 = document.setAttribute("id", "list-item")
+
+                // // Add text to link
+                li1.innerHTML ='<a href=' +YoutubeLink + '>Watch the trailer video: '+ trailerMovieName + '</a>'
+
+                // Append list items to ordered trailerListEl
+                trailerListEl.appendChild(li1);
+                
+                // Update Trailer element in HTML to have the new Youtube Link
+                // trailerSourceEl.setAttribute('src', YoutubeLink);
+                // trailerSourceEl.textContent = YoutubeLink;
+                // trailerContainerEl.textContent = YoutubeLink;
+            });   
+
+            } else {
+                // linkTrailerEl.css("display", "none");
+                console.log('getVideos API call not working');
+            }
+    })
+}    
 
 	// Hide the quiz questions when results are displayed
 	function hideQuiz() {
